feedbot-app/requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==3.0
decorator==4.4.2
feedparser==5.2.1
Flask==1.1.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.7.22
httpcore==0.9.1
httptools==0.1.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycparser==2.20
pydantic==1.6.1
python-telegram-bot==12.8
redis==3.5.3
rfc3986==1.4.0
six==1.15.0
sniffio==1.1.0
soupsieve==2.0.1
starlette==0.13.6
tornado==6.0.4
uvloop==0.14.0
websockets==8.1
Werkzeug==1.0.1
